backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
langgraph==0.2.53
mcp==1.3.0
sse-starlette==2.1.3
python-dotenv==1.0.1
requests==2.32.3
loguru==0.7.2
retry==0.9.2
openpyxl==3.1.5
httpx==0.27.2
PyExecJS==1.5.1
